chore: split auth module into separate files
