+++ crates/miden-lib/src/account/auth/rpo_falcon_512.rs
@@ -1,0 +1,39 @@
+use miden_objects::account::{AccountComponent, StorageSlot};
+use miden_objects::crypto::dsa::rpo_falcon512::PublicKey;
+
+use crate::account::components::rpo_falcon_512_library;
+
+/// An [`AccountComponent`] implementing the RpoFalcon512 signature scheme for authentication of
+/// transactions.
+///
+/// It reexports the procedures from `miden::contracts::auth::basic`. When linking against this
+/// component, the `miden` library (i.e. [`MidenLib`](crate::MidenLib)) must be available to the
+/// assembler which is the case when using [`TransactionKernel::assembler()`][kasm]. The procedures
+/// of this component are:
+/// - `auth__tx_rpo_falcon512`, which can be used to verify a signature provided via the advice
+///   stack to authenticate a transaction.
+///
+/// This component supports all account types.
+///
+/// [kasm]: crate::transaction::TransactionKernel::assembler
+pub struct AuthRpoFalcon512 {
+    public_key: PublicKey,
+}
+
+impl AuthRpoFalcon512 {
+    /// Creates a new [`AuthRpoFalcon512`] component with the given `public_key`.
+    pub fn new(public_key: PublicKey) -> Self {
+        Self { public_key }
+    }
+}
+
+impl From<AuthRpoFalcon512> for AccountComponent {
+    fn from(falcon: AuthRpoFalcon512) -> Self {
+        AccountComponent::new(
+            rpo_falcon_512_library(),
+            vec![StorageSlot::Value(falcon.public_key.into())],
+        )
+        .expect("falcon component should satisfy the requirements of a valid account component")
+        .with_supports_all_types()
+    }
+}
